fix(output): don't output empty files if there is nothing to write
